# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cocoon API Proxy",
    "description": "OpenAI-compatible API proxy for the Cocoon decentralized AI network on TON blockchain. Drop-in replacement for OpenAI client libraries with per-token billing and TON cryptocurrency payments. Use when you need access to decentralized AI models (e.g. Qwen/Qwen3-32B) via a standard OpenAI-format API with crypto-native billing.",
    "version": "0.1.0",
    "contact": {
      "url": "https://github.com/knyazev741/cocoon-api-proxy"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/v1/chat/completions": {
      "post": {
        "operationId": "create_chat_completion",
        "summary": "Create a chat completion",
        "description": "Send a conversation to a Cocoon AI model and receive a response. Supports both streaming (SSE) and non-streaming modes. Use when you need to generate AI text responses from models like Qwen/Qwen3-32B. Requires a funded API key — the cost is deducted per-token after the response completes.",
        "tags": ["Chat"],
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCompletionRequest"
              },
              "examples": {
                "simple": {
                  "summary": "Simple chat request",
                  "value": {
                    "model": "Qwen/Qwen3-32B",
                    "messages": [{"role": "user", "content": "Hello!"}]
                  }
                },
                "streaming": {
                  "summary": "Streaming chat request",
                  "value": {
                    "model": "Qwen/Qwen3-32B",
                    "messages": [{"role": "user", "content": "Explain quantum computing"}],
                    "stream": true,
                    "max_tokens": 1024
                  }
                },
                "system_prompt": {
                  "summary": "With system prompt",
                  "value": {
                    "model": "Qwen/Qwen3-32B",
                    "messages": [
                      {"role": "system", "content": "You are a helpful coding assistant."},
                      {"role": "user", "content": "Write a Python function to sort a list."}
                    ],
                    "temperature": 0.7
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat completion response with generated text and token usage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatCompletionResponse"
                },
                "example": {
                  "id": "chatcmpl-abc123",
                  "object": "chat.completion",
                  "created": 1711680000,
                  "model": "Qwen/Qwen3-32B",
                  "choices": [
                    {
                      "index": 0,
                      "message": {
                        "role": "assistant",
                        "content": "Hello! How can I help you today?"
                      },
                      "finish_reason": "stop"
                    }
                  ],
                  "usage": {
                    "prompt_tokens": 10,
                    "completion_tokens": 8,
                    "total_tokens": 18
                  }
                }
              },
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "description": "Server-Sent Events stream. Each event is a JSON object prefixed with 'data: '. Stream ends with 'data: [DONE]'."
                }
              }
            }
          },
          "400": {
            "description": "Invalid request — missing model or messages field",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "message": "model and messages are required",
                    "type": "invalid_request_error",
                    "param": null,
                    "code": null
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Insufficient balance — deposit more TON to continue",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "message": "Insufficient balance: 0.0050 TON. Minimum: 0.01 TON. Top up via /v1/deposit.",
                    "type": "insufficient_balance",
                    "param": null,
                    "code": "insufficient_balance"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/models": {
      "get": {
        "operationId": "list_models",
        "summary": "List available AI models",
        "description": "Returns the list of AI models available on the Cocoon network. Use to discover which model IDs are valid for chat completion requests. No authentication required.",
        "tags": ["Models"],
        "responses": {
          "200": {
            "description": "List of available models",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsResponse"
                },
                "example": {
                  "object": "list",
                  "data": [
                    {
                      "id": "Qwen/Qwen3-32B",
                      "object": "model",
                      "owned_by": "cocoon"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/keys": {
      "post": {
        "operationId": "create_api_key",
        "summary": "Create a new API key",
        "description": "Generate a new API key and deposit code for a new user account. The API key (ck_...) is returned once and cannot be retrieved again — save it immediately. The deposit code is used as a memo when sending TON to fund the account. If COCOON_ADMIN_TOKEN is configured, requires admin Bearer token.",
        "tags": ["Auth"],
        "security": [{"BearerAuth": []}],
        "responses": {
          "201": {
            "description": "New API key created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateKeyResponse"
                },
                "example": {
                  "api_key": "ck_a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2",
                  "deposit_code": "xK9mP2qR",
                  "balance_ton": 0.0,
                  "created_at": "2026-03-23 12:00:00"
                }
              }
            }
          },
          "403": {
            "description": "Admin token required but not provided or incorrect",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/balance": {
      "get": {
        "operationId": "get_balance",
        "summary": "Check account balance",
        "description": "Returns the current TON balance for the authenticated API key. Use to check available funds before making chat completion requests. Balance is shown in both TON and nanoton units.",
        "tags": ["Billing"],
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "Current balance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceResponse"
                },
                "example": {
                  "balance_ton": 1.5,
                  "balance_nanoton": 1500000000
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/deposit": {
      "get": {
        "operationId": "get_deposit_info",
        "summary": "Get deposit instructions",
        "description": "Returns the TON wallet address and unique memo code for depositing funds. Send TON to the wallet with the memo as the transaction comment — the balance is credited automatically within ~30 seconds. Use when you need to add funds to an API key account.",
        "tags": ["Billing"],
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "Deposit instructions with wallet address and memo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepositResponse"
                },
                "example": {
                  "wallet": "EQAv6GynfX6rrRlbNjYHpa5Mz1UDE61s5VPg-TN8sBFeR9Sp",
                  "memo": "xK9mP2qR",
                  "instructions": "Send TON to EQAv6GynfX6rrRlbNjYHpa5Mz1UDE61s5VPg-TN8sBFeR9Sp with the comment/memo: xK9mP2qR. Your balance will be credited automatically within ~30 seconds."
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "operationId": "health_check",
        "summary": "Health check",
        "description": "Returns service health status. Use to verify the API proxy is running and accepting requests. No authentication required.",
        "tags": ["System"],
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key authentication. Create a key via POST /v1/keys, then pass it as 'Authorization: Bearer ck_...'. Keys are prefixed with 'ck_' and are 67 characters long."
      }
    },
    "schemas": {
      "ChatMessage": {
        "type": "object",
        "required": ["role", "content"],
        "properties": {
          "role": {
            "type": "string",
            "enum": ["system", "user", "assistant"],
            "description": "The role of the message author"
          },
          "content": {
            "type": "string",
            "description": "The text content of the message"
          }
        }
      },
      "ChatCompletionRequest": {
        "type": "object",
        "required": ["model", "messages"],
        "properties": {
          "model": {
            "type": "string",
            "description": "Model ID to use. Get available models from GET /v1/models.",
            "example": "Qwen/Qwen3-32B"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "description": "Conversation messages. Must contain at least one message with role 'user'."
          },
          "stream": {
            "type": "boolean",
            "default": false,
            "description": "If true, returns Server-Sent Events stream. Each event contains a partial response chunk."
          },
          "max_tokens": {
            "type": "integer",
            "description": "Maximum tokens to generate. Capped at server-configured limit (default 128000).",
            "maximum": 128000
          },
          "temperature": {
            "type": "number",
            "description": "Sampling temperature between 0 and 2. Higher values produce more random output.",
            "minimum": 0,
            "maximum": 2
          },
          "stream_options": {
            "type": "object",
            "properties": {
              "include_usage": {
                "type": "boolean",
                "description": "If true, includes token usage in the final stream chunk"
              }
            }
          }
        }
      },
      "ChatCompletionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique completion identifier"
          },
          "object": {
            "type": "string",
            "enum": ["chat.completion"],
            "description": "Object type, always 'chat.completion'"
          },
          "created": {
            "type": "integer",
            "description": "Unix timestamp of creation"
          },
          "model": {
            "type": "string",
            "description": "Model used for completion"
          },
          "choices": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "index": {
                  "type": "integer"
                },
                "message": {
                  "$ref": "#/components/schemas/ChatMessage"
                },
                "finish_reason": {
                  "type": "string",
                  "enum": ["stop", "length", "content_filter"],
                  "description": "Why the model stopped generating"
                }
              }
            }
          },
          "usage": {
            "$ref": "#/components/schemas/UsageInfo"
          }
        }
      },
      "UsageInfo": {
        "type": "object",
        "properties": {
          "prompt_tokens": {
            "type": "integer",
            "description": "Tokens in the input prompt"
          },
          "completion_tokens": {
            "type": "integer",
            "description": "Tokens in the generated response"
          },
          "total_tokens": {
            "type": "integer",
            "description": "Total tokens (prompt + completion)"
          }
        }
      },
      "ModelsResponse": {
        "type": "object",
        "properties": {
          "object": {
            "type": "string",
            "enum": ["list"]
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "description": "Model identifier to use in chat completion requests"
                },
                "object": {
                  "type": "string",
                  "enum": ["model"]
                },
                "owned_by": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "CreateKeyResponse": {
        "type": "object",
        "properties": {
          "api_key": {
            "type": "string",
            "description": "The API key. Shown once — save it immediately. Format: ck_<64 hex chars>"
          },
          "deposit_code": {
            "type": "string",
            "description": "Unique 8-character code to use as memo when depositing TON"
          },
          "balance_ton": {
            "type": "number",
            "description": "Initial balance in TON (always 0 for new keys)"
          },
          "created_at": {
            "type": "string",
            "description": "ISO timestamp of key creation"
          }
        }
      },
      "BalanceResponse": {
        "type": "object",
        "properties": {
          "balance_ton": {
            "type": "number",
            "description": "Balance in TON (human-readable)"
          },
          "balance_nanoton": {
            "type": "integer",
            "description": "Balance in nanoton (1 TON = 1e9 nanoton)"
          }
        }
      },
      "DepositResponse": {
        "type": "object",
        "properties": {
          "wallet": {
            "type": "string",
            "description": "TON wallet address to send deposits to"
          },
          "memo": {
            "type": "string",
            "description": "Unique memo code to include as transaction comment"
          },
          "instructions": {
            "type": "string",
            "description": "Human-readable deposit instructions"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "description": "Human-readable error description"
              },
              "type": {
                "type": "string",
                "enum": ["invalid_request_error", "invalid_auth", "insufficient_balance", "auth_error", "server_error"],
                "description": "Error category"
              },
              "param": {
                "type": "string",
                "nullable": true,
                "description": "Parameter that caused the error, if applicable"
              },
              "code": {
                "type": "string",
                "nullable": true,
                "description": "Machine-readable error code: invalid_api_key, forbidden, insufficient_balance, internal_error"
              }
            }
          }
        }
      }
    }
  }
}